bot/slack-processor-python/pyproject.toml
[project]
name = "slack-bot-processor"
version = "0.1.0"
description = "Self-hosted Slack Bot for file processing pipelines"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "slack-bolt>=1.18.0",
    "python-dotenv>=1.0.0",
    "pillow>=10.0.0",
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.11"
strict = true
